requirements.txt
# Analizador de Artículos Científicos V2
# Dependencias del proyecto

# OpenAI API
openai>=1.0.0

# Procesamiento de PDF
PyMuPDF>=1.23.0

# Variables de entorno
python-dotenv>=1.0.0

# Interfaz gráfica (incluida en Python estándar)
# tkinter viene con Python por defecto

# Opcional: Para mejores logs
colorlog>=6.7.0
